feat(Status): merge profile concept into status
Previously, there was a Profile route that was used to display's a
user's current status and previous 4 statuses. The Profile concept was
to be extended to include followers and following, however as this is
now in the `/follow` routes it makes little sense to keep the Profile
concept around.

The Status concept is a logical place to retrieve user statuses
(unauthenticated) and there is a new route `GET /status/:username` that
will display any user's profile status.
