+++ web/controllers/actions/status_profile.ex
@@ -1,15 +1,15 @@
-defmodule StatazApi.ProfileController do
+defmodule StatazApi.StatusController.ActionProfile do
   use StatazApi.Web, :controller
   alias StatazApi.User
-  alias StatazApi.Profile
+  alias StatazApi.Status
 
-  def show(conn, %{"username" => username}) do
+  def execute(conn, %{"username" => username}) do
     Repo.get_by(User, %{username: username})
     |> response(conn)
   end
 
   defp response(user = %User{}, conn) do
-    profile = Profile.by_user_id(user.id)
+    profile = Status.profile_by_user_id(user.id, 5)
               |> Repo.all()
     render(conn, "show.json", profile: profile)
   end
