+++ web/views/status_view.ex
@@ -1,12 +1,16 @@
 defmodule StatazApi.StatusView do
   use StatazApi.Web, :view
 
+  def render("list.json", %{status: status}) do
+    %{data: render_many(status, StatazApi.StatusView, "status.json")}
+  end
+
   def render("show.json", %{status: status}) do
     %{data: render_one(status, StatazApi.StatusView, "status.json")}
   end
 
-  def render("list.json", %{status: status}) do
-    %{data: render_many(status, StatazApi.StatusView, "status.json")}
+  def render("show.json", %{profile: profile}) do
+    %{data: render_many(profile, StatazApi.StatusView, "profile.json")}
   end
 
   def render("show.json", %{error: error}) do
@@ -19,5 +23,11 @@ defmodule StatazApi.StatusView do
     %{id: status.id,
       description: status.description,
       active: status.active}
+  end
+
+  def render("profile.json", %{status: profile}) do
+    %{username: profile.user.username,
+      status: profile.description,
+      since: StatazApi.Util.Time.ecto_datetime_simple_format(profile.inserted_at)}
   end
 end
